+++ src/common_clj/state_flow_helpers/http.clj
@@ -1,0 +1,12 @@
+(ns common-clj.state-flow-helpers.http
+  (:require [common-clj.json :as json]
+            [io.pedestal.test :as test]
+            [state-flow.state :as state]))
+
+(defn request-arrived! [method path]
+  (state/gets
+   (fn [world]
+     (let [service-fn (-> world :system :http-server :service :io.pedestal.http/service-fn)
+           {:keys [status body]} (test/response-for service-fn method path)]
+       {:status status
+        :body   (json/string->json body)}))))